# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Road application
  title: Платные дороги
  version: "1.0"
paths:
...
  /travelrequestroads/{requestID}/{roadID}:
    delete:
      consumes:
      - application/json
      description: delete road from travelrequest
      operationId: delete-road-from-travelrequest
      parameters:
      - description: ID заявки
        in: path
        name: requestID
        required: true
        type: integer
      - description: ID консультации
        in: path
        name: roadID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Консультация была удалена из заявки
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Road From TravelRequest
      tags:
      - Road-TravelRequest
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"